requirements.txt
darkskylib
discord.py[voice]
youtube_dl
aiohttp
wikipedia
python-forecastio
geocoder
pyping
youtube_dl
postgres
asyncqlio
colorlog
psutil
random-cat
nmap
pythonwhois
retry